requirements.txt
# Core ML
sentence-transformers>=2.7.0
xgboost>=2.1.0
scikit-learn>=1.6.0

# Data Processing
pdfplumber>=0.10.0
pytesseract>=0.3.10
pdf2image>=1.16.0
pandas>=2.2.0
numpy>=2.1.0

# NER & LLM
google-genai>=1.0.0

# UI Framework
streamlit>=1.32.0
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Evaluation & Statistics
scipy>=1.14.0

# Utilities
openpyxl>=3.1.0
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.66.0
